AI-Based-Career-Recommendation-System/apps/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pip install "pydantic[email]"

